♻️ Refactor(shared) : [코드리뷰 반영] useState 해당하는 컴포넌트로 이동
- App.js에서 전부 선언했던 useState를 사용하는 컴포넌트 안으로 이동하였습니다. -> 리렌더링 고려
