+++ src/components/shared/Header.js
@@ -1,0 +1,28 @@
+import '../../styles/header.css';
+import { useEffect, useState } from 'react';
+import { getData } from '../../utils/getData';
+import PATH from '../../constants/path';
+import Profile from '../Profile';
+import LoginBtn from '../LoginBtn';
+import logoImg from '../../assets/images/logo.svg';
+
+function Header() {
+  const [user, setUser] = useState(null);
+
+  useEffect(() => {
+    getData(setUser, PATH.user);
+  }, []);
+
+  return (
+    <header>
+      <div className="header_container">
+        <a href="/" target="_blank" rel="noopener noreferrer">
+          <img className="linkbrary_logo" src={logoImg} alt="Linkbrary 메인 페이지 바로가기" />
+        </a>
+        {user ? <Profile user={user} /> : <LoginBtn />}
+      </div>
+    </header>
+  );
+}
+
+export default Header;
