+++ src/styles/folderInfo.css
@@ -1,16 +1,11 @@
 .folder_info {
-  background-color: var(--bg);
-  width: 100%;
-  padding: 80px 0px 60px;
-}
-
-.folder_info_container {
-  margin: 0 auto;
-  width: 210px;
   display: flex;
   flex-direction: column;
   align-items: center;
   gap: 20px;
+  background-color: var(--bg);
+  width: 100%;
+  padding: 120px 0px 60px;
   text-align: center;
 }
